refactor: move common utils to shared file
