+++ src/tigerflow_ml/text/chat_completion/_base.py
@@ -10,8 +10,7 @@ from tigerflow.logconfig import logger
 from tigerflow.utils import SetupContext
 
 from tigerflow_ml.params import VLLMParams
-
-from .utils import EmptyFileError, parse_kwargs, read_file_with_fallback
+from tigerflow_ml.utils import EmptyFileError, parse_kwargs, read_file_with_fallback
 
 _TEXT_EXTENSIONS = [".txt", ".text", ".md", ".log", ".rtf"]
 _IMG_EXTENSIONS = [".jpg", ".jpeg", ".png", ".tiff", ".tif", ".bmp"]
