+++ src/tigerflow_ml/text/translate/_base.py
@@ -23,6 +23,12 @@ from tigerflow.logconfig import logger
 from tigerflow.utils import SetupContext
 
 from tigerflow_ml.params import VLLMParams
+from tigerflow_ml.utils import (
+    ConfigParsingError,
+    EmptyFileError,
+    parse_kwargs,
+    read_file_with_fallback,
+)
 
 from .chunking import (
     DEFAULT_CHUNK_SIZE,
@@ -32,13 +38,9 @@ from .chunking import (
     count_tokens,
 )
 from .detection import LANGUAGES, detect_language, get_language_name
-from .utils import (
+from .errors import (
     AlreadyInTargetLanguageError,
-    ConfigParsingError,
-    EmptyFileError,
     TranslationError,
-    parse_kwargs,
-    read_file_with_fallback,
 )
 
 if TYPE_CHECKING:
